cli: refactor config to the new approach
